(refactor) detail directory
