+++ detail.html
@@ -20,11 +20,7 @@
     <link rel="stylesheet" href="css/detail/providers.css" />
     <link rel="stylesheet" href="css/detail/similar-movie.css" />
     <!-- JS 가져오기 -->
-
     <script type="module" src="js/detail/index.js"></script>
-    <script type="module" src="js/detail/localStorage.js"></script>
-    <script type="module" src="./js/detail/icon.js"></script>
-    <script type="module" src="js/detail/detailDisplay.js"></script>
   </head>
   <body>
     <!-- 헤더 -->
