+++ js/detail/index.js
@@ -1,9 +1,11 @@
 import useMovieDetailData from "./hook/useMovieDetailData.js";
-import displayActors from "./displayActors.js";
-import displayProviders from "./displayProviders.js";
-import displayTrailer from "./displayTrailer.js";
-import displaySimilarMovies from "./displaySimilarMovies.js";
 import addSearchEvent from "../search/addSearchEvent.js";
+import displayActors from "./display/displayActors.js";
+import displayProviders from "./display/displayProviders.js";
+import displayTrailer from "./display/displayTrailer.js";
+import displaySimilarMovies from "./display/displaySimilarMovies.js";
+import displayMovieDetail from "./display/displayMovieDetail.js";
+import initReviews from "./review/initReviews.js";
 
 const {
   getActorsData,
@@ -31,6 +33,8 @@ const providerElements = displayProviders(providers);
 const trailerElement = displayTrailer(trailer);
 const similarMovieElements = displaySimilarMovies(similarMovies);
 
+displayMovieDetail();
+
 actorsWrap.innerHTML = actorElements;
 if (providerElements) {
   providersWrap.innerHTML = providerElements;
@@ -39,3 +43,5 @@ trailerWrap.innerHTML = trailerElement;
 similarMoviesWrap.innerHTML = similarMovieElements;
 
 addSearchEvent();
+
+initReviews();
