+++ js/detail/review/initReviews.js
@@ -4,8 +4,10 @@ import { handleDelete } from "./handleDelete.js";
 import { displayAllReviews } from "./displayAllReviews.js";
 import { handleEdit } from "./handleEdit.js";
 
-reviewForm.addEventListener("submit", handleSubmit);
-reviewContainer.addEventListener("click", handleDelete);
-reviewContainer.addEventListener("click", handleEdit);
+export default function initReviews() {
+  reviewForm.addEventListener("submit", handleSubmit);
+  reviewContainer.addEventListener("click", handleDelete);
+  reviewContainer.addEventListener("click", handleEdit);
 
-displayAllReviews();
+  displayAllReviews();
+}
